# Enable use of different Jacobian implementations with 1D solver