--- a/include/cantera/numerics/AdaptivePreconditioner.h
+++ b/include/cantera/numerics/AdaptivePreconditioner.h
@@ -45,6 +45,10 @@ public:
 
     void updatePreconditioner() override;
 
+    int info() const override {
+        return static_cast<int>(m_solver.info());
+    }
+
     //! Prune preconditioner elements
     void prunePreconditioner();