--- a/include/cantera/numerics/PreconditionerFactory.h
+++ b/include/cantera/numerics/PreconditionerFactory.h
@@ -3,34 +3,13 @@
 // This file is part of Cantera. See License.txt in the top-level directory or
 // at https://cantera.org/license.txt for license and copyright information.
 
-#ifndef PRECONDITIONER_FACTORY_H
-#define PRECONDITIONER_FACTORY_H
+#ifndef PRECONDITIONERFACTORY_FACTORY_H
+#define PRECONDITIONERFACTORY_FACTORY_H
 
-#include "cantera/base/FactoryBase.h"
+#pragma message("warning: PreconditionerFactory.h and class PreconditionerFactory " \
+                "are deprecated and will be removed after Cantera 3.2. Use " \
+                "SystemJacobianFactory.h and class SystemJacobianFactory.")
 
-namespace Cantera
-{
-
-class PreconditionerBase;
-
-//! Factory class to create preconditioner objects
-class PreconditionerFactory : public Factory<PreconditionerBase>
-{
-public:
-    static PreconditionerFactory* factory();
-
-    //! Delete preconditioner factory
-    void deleteFactory() override;
-
-private:
-    static PreconditionerFactory* s_factory;
-    static std::mutex precon_mutex;
-    PreconditionerFactory();
-};
-
-//! Create a Preconditioner object of the specified type
-shared_ptr<PreconditionerBase> newPreconditioner(const string& precon);
-
-}
+#include "SystemJacobianFactory.h"
 
 #endif