… doc/sphinx/python/index.md
… doc/sphinx/python/utilities.rst
… doc/sphinx/python/zerodim.rst
… include/cantera/oneD/MultiNewton.h
… include/cantera/oneD/OneDim.h
… include/cantera/zeroD/ReactorNet.h
… interfaces/cython/cantera/__init__.pxd
… interfaces/cython/cantera/_cantera.pyx
… interfaces/cython/cantera/_onedim.pxd
… interfaces/cython/cantera/_onedim.pyx
… interfaces/cython/cantera/jacobians.pxd
… interfaces/cython/cantera/jacobians.pyx
… interfaces/cython/cantera/preconditioners.pxd
… interfaces/cython/cantera/preconditioners.pyx
… interfaces/cython/cantera/reactor.pxd
… interfaces/cython/cantera/reactor.pyx
… src/numerics/AdaptivePreconditioner.cpp
… src/numerics/EigenSparseDirectJacobian.cpp
… src/numerics/EigenSparseJacobian.cpp
… src/numerics/PreconditionerFactory.cpp
… src/numerics/SystemJacobianFactory.cpp
… src/oneD/Domain1D.cpp
… src/oneD/MultiJac.cpp
… src/oneD/MultiNewton.cpp
… src/oneD/OneDim.cpp
… src/oneD/Sim1D.cpp
… src/zeroD/ReactorNet.cpp
… test/python/test_onedim.py
… test/zeroD/test_zeroD.cpp
… test_problems/SConscript