<!-- Thanks for contributing code! Please include a description of your change and check your pull request against the list below. For further questions, refer to the contributing guide (https://github.com/Cantera/cantera/blob/main/CONTRIBUTING.md). -->

**Changes proposed in this pull request**

<!-- Provide a clear and concise description of changes and/or features introduced in this pull request. -->

- Generalize `PreconditionerBase` (now renamed `SystemJacobian`) to be able represent both preconditioners ($M = I - \gamma J$) and the matrices needed for the pseudo-transient solver used by `Sim1D` ($M = J - \alpha \Delta t^{-1}$)
- Refactor the `MultiJac` to be a derived class of `PreconditionerBase` and avoid direct inheritance from `BandMatrix`
- Add methods to `OneDim` to allow setting the implementation of `SystemJacobian` to use
- Introduce a `SystemJacobian` implementation that uses Eigen's sparse direct LU factorization
- Improve selection of perturbations used when computing the finite difference Jacobian used by `OneDim`. The new perturbation sizes are based on those used by SUNDIALS, and generally improve the quality of the Jacobian (reducing the number of residual and Jacobian evaluations required to converge).

**If applicable, fill in the issue number this pull request is fixing**

This is preliminary work toward several planned improvements:
- Cantera/enhancements#31, where the banded Jacobian is a poor default (related: #1021, #989, #654)
- Cantera/enhancements#86
- Cantera/enhancements#34
- Cantera/enhancements#61 (related: #609)

<!-- Issues with issue number '<issue>' are referenced as #<issue>. To link to an issue in the enhancements repository, use Cantera/enhancements#<issue>. -->

**If applicable, provide an example illustrating new features this pull request is introducing**

Besides supporting these other improvements, using the sparse LU solver provided by Eigen along with the improved Jacobian perturbations provides some significant speedups to the 1D solver. Using the Blanquart et al (2018) mechanism from [here](https://shepherd.caltech.edu/EDL/PublicResources/sdt/cti_mech.html) (194 species, 1156 reactions), I get the following timings for a laminar flame speed calculation:
- Windows, using SUNDIALS to provide the banded solver:
  - baseline: 154.9 s
  - new Jacobian perturbations: 104.1 s (-34%)
  - sparse: 84.3 s (-46%)
- Linux, using MKL to provide the banded solver:
  - baseline (32 processors): 210.3 s wall time, 2064 s user CPU
  - banded, new Jacobian perturbations (1 processor): 190.4 s wall time (-9.4%)
  - sparse: 157.6 s wall time (-25%)

**Checklist**

- [x] The pull request includes a clear description of this code change
- [x] Commit messages have short titles and reference relevant issues
- [ ] Build passes (`scons build` & `scons test`) and unit tests address code coverage
- [x] Style & formatting of contributed code follows [contributing guidelines](https://github.com/Cantera/cantera/blob/main/CONTRIBUTING.md)
- [ ] The pull request is ready for review